% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_to_01.R
\name{bin}
\alias{bin}
\title{Binomial function}
\usage{
bin(n, i, x)
}
\arguments{
\item{n}{Integer.}

\item{i}{Integer(i < n).}

\item{x}{Numeric(0<= x <=1).}
}
\description{
Binomial function
}
\examples{
bin(5,3,.5)
}
