% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BP2Dcoef.R
\name{BP2D_coef}
\alias{BP2D_coef}
\title{Getting coefficeint from BP2D result.}
\usage{
BP2D_coef(result)
}
\arguments{
\item{result}{This is output of BP2D.}
}
\value{
Coefficients table.
}
\description{
This function will return coefficient and length of each set of coefficeint.
}
\seealso{
Other Bayesain estimate: 
\code{\link{BP2D_table}()},
\code{\link{BP2D}()}
}
\concept{Bayesain estimate}
