% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BP2Dtable.R
\name{BP2D_table}
\alias{BP2D_table}
\title{Table and Criterion.}
\usage{
BP2D_table(results_list)
}
\arguments{
\item{results_list}{A vector of characters.}
}
\value{
Table and criterion T.
}
\description{
If you give more groups of prior,you can use this function to get the table and T criterion.
}
\seealso{
Other Bayesain estimate: 
\code{\link{BP2D_coef}()},
\code{\link{BP2D}()}
}
\concept{Bayesain estimate}
