% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPbasis.R
\name{BPbasis}
\alias{BPbasis}
\title{Bernstein polynomial basis.}
\usage{
BPbasis(n0, ages, years, nn = 1)
}
\arguments{
\item{n0}{Upper bound of possion random variable.}

\item{ages}{Range of ages.}

\item{years}{Range of years.}

\item{nn}{Lower bound of possion random variable.}
}
\value{
Bernstein basis.
}
\description{
This function build two dimensional Bernstein polynomial basis.
}
\examples{
ages<-35:85
years<-1988:2007
list.basis<-BPbasis(10,ages,years)
list.basis
}
\seealso{
Other Bernstein basis: \code{\link{BPbasis_pd_ages}},
  \code{\link{BPbasis_pd_years}}
}
\concept{Bernstein basis}
