% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPbasis_pd_ages.R
\name{BPbasis_pd_ages}
\alias{BPbasis_pd_ages}
\title{Partial differential Bernstein polynomial basis.}
\usage{
BPbasis_pd_ages(n0, ages, years, nn = 1)
}
\arguments{
\item{n0}{Upper bound of possion random variable.}

\item{ages}{Range of ages.}

\item{years}{Range of years.}

\item{nn}{Lower bound of possion random variable.}
}
\value{
Partial differential Bernstein basis by ages.
}
\description{
This function build two dimensional partial differential Bernstein polynomial basis by ages.
}
\examples{
ages<-35:85
years<-1988:2007
list.basis<-BPbasis_pd_ages(10,ages,years)
list.basis
}
\seealso{
Other Bernstein basis: \code{\link{BPbasis_pd_years}},
  \code{\link{BPbasis}}
}
\concept{Bernstein basis}
