% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPFhat_pd_years.R
\name{BPFhat_pd_years}
\alias{BPFhat_pd_years}
\title{Bernstein polynomial}
\usage{
BPFhat_pd_years(coef, ages, years)
}
\arguments{
\item{coef}{Bernstein polynimial coefficients.}

\item{ages}{Range of ages.}

\item{years}{Range of years.}
}
\value{
This function return outer Bernstein polynomial using coefficients.
}
\description{
Given Bernstein polynomial coeffients to compute partial differential by years Fhat.
}
\examples{
coef<-runif(9)
ages<-35:85
years<-1988:2007
BPFhat_pd_years(coef,ages,years)
}
\seealso{
Other outer Bernstein polynomial: \code{\link{BPFhat_pd_ages}},
  \code{\link{BPFhat}}
}
\concept{outer Bernstein polynomial}
