% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Credible_interval.R
\name{Credible_interval}
\alias{Credible_interval}
\title{Credible interval.}
\usage{
Credible_interval(result, n_cluster = 1, alpha = 0.05)
}
\arguments{
\item{result}{This is output of BP2D.}

\item{n_cluster}{Muticores is remmended.(default:n_cluster=1)}

\item{alpha}{Level of significance.}
}
\value{
Bayesian credible interval with level of significance.
}
\description{
Builing two dimensional Bernstein polynomial credible interval.
}
\references{
L.H. Chien, T.J. Tseng, C.H. Chen, H.F. Jiang, F.Y. Tsai, T.W. Liu, C.A. Hsiung, I.S. Chang Comparison of annual percentage change in breast cancer incidence rate between Taiwan and the United States-A smoothed Lexis diagram approach.
}
\seealso{
Other Credible interval: 
\code{\link{PD_Credible_interval}()}
}
\concept{Credible interval}
