% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Credible_interval_pd_ages.R
\name{Credible_interval_pd_ages}
\alias{Credible_interval_pd_ages}
\title{Credible interval.}
\usage{
Credible_interval_pd_ages(result, n_cluster = 1, alpha = 0.05)
}
\arguments{
\item{result}{This is output of BP2D.}

\item{n_cluster}{Muticores is remmended.(default:n_cluster=1)}

\item{alpha}{Level of significance.}
}
\value{
This function return Bayesian credible interval with level of significance.
}
\description{
Builing partial differential Bernstein polynomial credible interval.
}
\seealso{
Other Credible interval: \code{\link{Credible_interval_pd_years}},
  \code{\link{Credible_interval}}
}
\concept{Credible interval}
