% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.R
\name{gen_data}
\alias{gen_data}
\title{Generated data}
\usage{
gen_data(ages, years, FT, M)
}
\arguments{
\item{ages}{Ages.}

\item{years}{Years.}

\item{FT}{Rate function.}

\item{M}{Population function.}
}
\description{
Generated data
}
\keyword{datasets}
