% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_to_01.R
\name{mapping_to_01}
\alias{mapping_to_01}
\title{Mapping to [0,1]}
\usage{
mapping_to_01(x)
}
\arguments{
\item{x}{Vector.}
}
\description{
Mapping to [0,1]
}
\examples{
mapping_to_01(35:85)
(35:85)/(85-35)
mapping_to_01(runif(10))
}
