% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_BPbasis.R
\name{PD_BPbasis}
\alias{PD_BPbasis}
\title{Partial differential Bernstein polynomial basis.}
\usage{
PD_BPbasis(ages, years, n0, N = 1, by = 1)
}
\arguments{
\item{ages}{Range of ages.}

\item{years}{Range of years.}

\item{n0}{Upper bound of possion random variable.}

\item{N}{Lower bound of possion random variable.}

\item{by}{1: partial differential by ages; 2: partial differential by years.}
}
\value{
Partial differential Bernstein basis.
}
\description{
This function build two dimensional Bernstein polynomial basis.
}
\examples{
ages <- 35:85
years <- 1988:2007
pdbasis <- PD_BPbasis(ages,years,10,by = 1)
pdbasis
}
\seealso{
Other Bernstein basis: 
\code{\link{BPbasis}()}
}
\concept{Bernstein basis}
