% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_to_01.R
\name{scale_to_01}
\alias{scale_to_01}
\title{Scale to [0,1]}
\usage{
scale_to_01(x)
}
\arguments{
\item{x}{Vector.}
}
\description{
Scale to [0,1]
}
\examples{
scale_to_01(35:85)
(35:85-35)/(85-35)
scale_to_01(runif(10))
}
