% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.R
\docType{data}
\name{simulated_data_2}
\alias{simulated_data_2}
\title{Generate simulated data 2}
\format{
list of matrix
}
\usage{
data(simulated_data_2)
}
\description{
Given rate function 2 generated data.
}
\examples{
ages <- 35:85
years <- 1988:2007
FT2 <- function(x,y){0.00148*sin(0.5*pi*x*(y+0.2))+0.00002}
simulated_data_2 <- gen_data(ages,years,FT2,M)
}
\keyword{datasets}
