% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBP.R
\name{write.BP}
\alias{write.BP}
\title{Write xlsx file}
\usage{
write.BP(writedata, filename)
}
\arguments{
\item{writedata}{result of BP2D(character or list).}

\item{filename}{xlsx file name.}
}
\description{
This function will write result of BP2D to xlsx file.
}
