% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBPtable.R
\name{write.BPtable}
\alias{write.BPtable}
\title{Write BPtalbe as xlsx file}
\usage{
write.BPtable(BPtable, filename)
}
\arguments{
\item{BPtable}{output of BP2D_table.}

\item{filename}{xlsx file name.}
}
\description{
If your environment has some result of BP2D,then you can use this function to store BPTable.
}
