% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPbasis.R
\name{BPbasis}
\alias{BPbasis}
\title{Bernstein polynomial basis.}
\usage{
BPbasis(ages, years, n0, N = 1)
}
\arguments{
\item{ages}{Range of ages.}

\item{years}{Range of years.}

\item{n0}{Upper bound of possion random variable.}

\item{N}{Lower bound of possion random variable.}
}
\value{
Bernstein basis.
}
\description{
This function build two dimensional Bernstein polynomial basis.
}
\examples{
ages <- 35:85
years <- 1988:2007
list.basis <- BPbasis(ages,years,10)
list.basis
}
\seealso{
Other Bernstein basis: 
\code{\link{PD_BPbasis}()}
}
\concept{Bernstein basis}
