% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPFhat.R
\name{BPFhat}
\alias{BPFhat}
\title{Two dimensional Bernstein polynomial}
\usage{
BPFhat(coef, ages, years, basis)
}
\arguments{
\item{coef}{Bernstein polynomial coefficients.}

\item{ages}{Range of ages.}

\item{years}{Range of years.}

\item{basis}{Bernstein polynomial basis.}
}
\value{
This function return outer Bernstein polynomial using coefficients.
}
\description{
Given Bernstein polynomial coeffients to compute Fhat.
}
\examples{
coef <- runif(9)
ages <- 35:85
years <- 1988:2007
list.basis <- BPbasis(ages,years,10)
BPFhat(coef,ages,years,list.basis)
}
\seealso{
Other outer Bernstein polynomial: 
\code{\link{PD_BPFhat}()}
}
\concept{outer Bernstein polynomial}
