% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.R
\docType{data}
\name{simulated_data_1}
\alias{simulated_data_1}
\title{Generate simulated data 1}
\format{
list of matrix
}
\usage{
data(simulated_data_1)
}
\description{
Given rate function 1 generated data.
}
\examples{
ages <- 35:85
years <- 1988:2007
FT1 <- function(x,y){0.00148*sin(0.5*pi*x*y)+0.00002}
simulated_data_1 <- gen_data(ages,years,FT1,M)
}
\keyword{datasets}
